Move page counter to thread parsing package
This resolves a circular import error.
